Extract command-line dispatch
