--- src/jlmkr/__about__.py
@@ -1,12 +1,0 @@
-# SPDX-FileCopyrightText: © 2024 Jip-Hop and the Jailmakers <https://github.com/Jip-Hop/jailmaker>
-#
-# SPDX-License-Identifier: LGPL-3.0-only
-
-__version__ = "3.0.0.dev1"
-
-__author__ = "Jip-Hop"
-__copyright__ = "Copyright © 2023, Jip-Hop and the Jailmakers"
-__license__ = "LGPL-3.0-only"
-
-__disclaimer__ = """USE THIS SCRIPT AT YOUR OWN RISK!
-IT COMES WITHOUT WARRANTY AND IS NOT SUPPORTED BY IXSYSTEMS."""
