--- src/jlmkr/donor/__init__.py
@@ -1,13 +1,0 @@
-# SPDX-FileCopyrightText: © 2024 Jip-Hop and the Jailmakers <https://github.com/Jip-Hop/jailmaker>
-#
-# SPDX-License-Identifier: LGPL-3.0-only
-
-from .jlmkr import main
-
-####
-#
-# Transitional, as we split up jlmkr.py into smaller components
-#
-# <https://github.com/Jip-Hop/jailmaker/discussions/231#discussioncomment-10024399>
-#
-####
