+++ src/jlmkr/utils/dataset.py
@@ -10,9 +10,10 @@ import subprocess
 
 from pathlib import PurePath
 from textwrap import dedent
+
+from paths import JAILS_DIR_PATH, JAIL_CONFIG_NAME, JAIL_ROOTFS_NAME, SCRIPT_DIR_PATH
 from utils.chroot import Chroot
 from utils.console import eprint, YELLOW, BOLD, NORMAL
-from utils.paths import JAILS_DIR_PATH, JAIL_CONFIG_NAME, JAIL_ROOTFS_NAME, SCRIPT_DIR_PATH
 
 
 def get_jail_path(jail_name):
