+++ src/jlmkr/utils/gpu.py
@@ -7,9 +7,10 @@ import subprocess
 
 from pathlib import Path
 from textwrap import dedent
+
+from paths import SHORTNAME
 from utils.console import eprint
 from utils.dataset import get_jail_rootfs_path
-from utils.paths import SHORTNAME
 
 
 # Test intel GPU by decoding mp4 file (output is discarded)
